backend/apis/game/connect_postgres.py
from configparser import ConfigParser
import psycopg2
import time

def load_config(filename='/usr/bin/database.ini', section='postgresql'):
    parser = ConfigParser()
    parser.read(filename)

    # get section, default to postgresql
    config = {}

    if parser.has_section(section):
        params = parser.items(section)
        for param in params:
            config[param[0]] = param[1]
    else:
        raise Exception('Section {0} not found in the {1} file'.format(section, filename))

    return config

def connect(config):
    """ Connect to the PostgreSQL database server """
    while True:
        print("Connecting to PostgreSQL database server...")
        try:
            # connecting to the PostgreSQL server
            with psycopg2.connect(**config) as conn:
                print('Connected to the PostgreSQL server.')
                return conn
        except (psycopg2.DatabaseError, Exception) as error:
            pass
            #print("Error trying to connect:", error)
        time.sleep(0.5)
        

config = load_config()
connect(config)