backend/apis/game/requirements.txt
setuptools>=58.0.0
#autobahn==23.1.2
wheel
asgiref==3.7.2
Django==4.2.11
sqlparse==0.4.4
psycopg2-binary
python-dotenv==0.10.1
requests
numpy
python-math
httpx
channels[daphne]
channels_redis